requirements.txt
langchain>=0.2.10
langchain-community>=0.2.9
langchain-openai>=0.1.17
langchain-text-splitters>=0.2.2
chromadb>=0.5.4
tiktoken>=0.7.0
pypdf>=4.2.0
docx2txt>=0.8
python-docx>=1.1.2
fpdf2>=2.7.9
pytest>=8.2.0
